Move `share.proto` to a separate package.
