+++ model/src/main/proto/spine_examples/shareaware/market/availbale_market_shares.proto
@@ -36,7 +36,7 @@ option java_outer_classname = "AvailableMarketSharesProto";
 option java_multiple_files = true;
 
 import "spine_examples/shareaware/identifiers.proto";
-import "spine_examples/shareaware/share.proto";
+import "spine_examples/shareaware/share/share.proto";
 
 // Displays the available shares on the market.
 message AvailableMarketShares {
@@ -46,5 +46,5 @@ message AvailableMarketShares {
     MarketId id = 1;
 
     // Available shares on the market.
-    repeated Share share = 2 [(required) = true];
+    repeated share.Share share = 2 [(required) = true];
 }
