+++ model/src/test/java/io/spine/examples/shareaware/sharesreader/SharesReaderTestEnv.java
@@ -27,7 +27,7 @@
 package io.spine.examples.shareaware.sharesreader;
 
 import com.google.common.collect.ImmutableSet;
-import io.spine.examples.shareaware.Share;
+import io.spine.examples.shareaware.share.Share;
 import io.spine.examples.shareaware.ShareId;
 
 import java.util.Set;
