+++ server/src/main/java/io/spine/examples/shareaware/server/market/MarketData.java
@@ -27,7 +27,7 @@
 package io.spine.examples.shareaware.server.market;
 
 import com.google.common.collect.ImmutableList;
-import io.spine.examples.shareaware.Share;
+import io.spine.examples.shareaware.share.Share;
 import io.spine.examples.shareaware.sharesreader.SharesReader;
 import io.spine.money.Money;
