+++ server/src/test/java/io/spine/examples/shareaware/server/market/MarketDataTest.java
@@ -26,7 +26,7 @@
 
 package io.spine.examples.shareaware.server.market;
 
-import io.spine.examples.shareaware.Share;
+import io.spine.examples.shareaware.share.Share;
 import io.spine.testing.UtilityClassTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
